Move dummy cache data generator into a separate file
